@@ -1,11 +1,16 @@
 <script lang="ts">
 	import type { LayoutData } from './$types';
 
-	export let data: LayoutData;
+	interface Props {
+		data: LayoutData;
+		children?: import('svelte').Snippet;
+	}
+
+	let { data, children }: Props = $props();
 </script>
 
 {#if data.isLoggedIn}
-	<slot />
+	{@render children?.()}
 {:else}
     <p>{data.translations.requiresLogIn}</p>
 {/if}
