@@ -4,7 +4,7 @@
 	import VybiratkoFirmy from '$lib/components/VybiratkoFirmy.svelte';
 	import Scanner from '$lib/components/Scanner.svelte';
 
-	import * as v from '$lib/Vec';
+	import * as v from '$lib/Vec.svelte';
 	import {
 		dataToRawData,
 		newData,
@@ -16,7 +16,7 @@
 	import { responsiblePerson } from '$lib/client/realtime';
 	import { companies, filteredCompanies } from './companies';
 	import odeslat from './odeslat';
-	import { p } from '$lib/Vec';
+	import { p } from '$lib/Vec.svelte';
 	import { evidence } from '$lib/client/firestore';
 	import { storableOrUndefined } from '$lib/helpers/stores';
 
@@ -30,8 +30,8 @@
 
 	const storedData = storableOrUndefined<RawData>(`stored_data`);
 
-	let mode: 'loading' | 'create' | 'edit' | 'createStored' = 'loading';
-	let data: Data = newData();
+	let mode: 'loading' | 'create' | 'edit' | 'createStored' = $state('loading');
+	let data: Data = $state(newData());
 	onMount(async () => {
 		const ir = $page.url.searchParams.get('edit');
 		if (ir) {
@@ -51,92 +51,114 @@
 		}
 	});
 
-	$: if (mode != 'loading') {
-		data.ostatni.zodpovednaOsoba.zobrazit = () => $responsiblePerson == null;
-		if ($responsiblePerson != null) data.ostatni.zodpovednaOsoba.value = $responsiblePerson;
-	}
-	$: if (mode != 'loading') {
-		if (data.uvedeni.jakoMontazka.value) {
-			data.uvedeni.ico.updateText('');
-			data.uvedeni.email.value = '';
-			data.uvedeni.phone.value = '';
-		} else if (
-			data.uvedeni.ico.value == data.montazka.ico.value &&
-			data.uvedeni.ico.value != '' &&
-			data.uvedeni.email.value == data.montazka.email.value &&
-			data.uvedeni.email.value != '' &&
-			data.uvedeni.phone.value == data.montazka.phone.value &&
-			data.uvedeni.phone.value != ''
-		) {
-			data.uvedeni.jakoMontazka.value = true;
-			data.uvedeni.ico.updateText('');
-			data.uvedeni.email.value = '';
-			data.uvedeni.phone.value = '';
+	$effect(() => {
+		data
+		if (mode != 'loading') {
+			data.ostatni.zodpovednaOsoba.zobrazit = () => $responsiblePerson == null;
+			if ($responsiblePerson != null) data.ostatni.zodpovednaOsoba.value = $responsiblePerson;
 		}
-	}
-	$: if (mode != 'loading') {
-		if (data.ir.typ.value.first == p`IR 12`) {
-			data.ir.typ.value.second = p`CTC`;
+	});
+	$effect(() => {
+		data
+		if (mode != 'loading') {
+			if (data.uvedeni.jakoMontazka.value) {
+				data.uvedeni.ico.updateText('');
+				data.uvedeni.email.value = '';
+				data.uvedeni.phone.value = '';
+			} else if (
+				data.uvedeni.ico.value == data.montazka.ico.value &&
+				data.uvedeni.ico.value != '' &&
+				data.uvedeni.email.value == data.montazka.email.value &&
+				data.uvedeni.email.value != '' &&
+				data.uvedeni.phone.value == data.montazka.phone.value &&
+				data.uvedeni.phone.value != ''
+			) {
+				data.uvedeni.jakoMontazka.value = true;
+				data.uvedeni.ico.updateText('');
+				data.uvedeni.email.value = '';
+				data.uvedeni.phone.value = '';
+			}
 		}
-	}
-	$: if (mode != 'loading') {
-		if (data.mistoRealizace.jakoBydliste.value) {
-			data.mistoRealizace.obec.updateText('');
-			data.mistoRealizace.psc.value = '';
-			data.mistoRealizace.ulice.value = '';
-		} else if (
-			data.mistoRealizace.obec.value == data.bydliste.obec.value &&
-			data.mistoRealizace.obec.value != '' &&
-			data.mistoRealizace.psc.value == data.bydliste.psc.value &&
-			data.mistoRealizace.psc.value != '' &&
-			data.mistoRealizace.ulice.value == data.bydliste.ulice.value &&
-			data.mistoRealizace.ulice.value != ''
-		) {
-			data.mistoRealizace.jakoBydliste.value = true;
-			data.mistoRealizace.obec.updateText('');
-			data.mistoRealizace.psc.value = '';
-			data.mistoRealizace.ulice.value = '';
+	});
+	$effect(() => {
+		data
+		if (mode != 'loading') {
+			if (data.ir.typ.value.first == p`IR 12`) {
+				data.ir.typ.value.second = p`CTC`;
+			}
 		}
-	}
-	$: if (mode != 'loading') {
-		if (!data.tc.model.moznosti(data).includes(data.tc.model.value ?? '')) {
-			data.tc.model.value = null;
+	});
+	$effect(() => {
+		data
+		if (mode != 'loading') {
+			if (data.mistoRealizace.jakoBydliste.value) {
+				data.mistoRealizace.obec.updateText('');
+				data.mistoRealizace.psc.value = '';
+				data.mistoRealizace.ulice.value = '';
+			} else if (
+				data.mistoRealizace.obec.value == data.bydliste.obec.value &&
+				data.mistoRealizace.obec.value != '' &&
+				data.mistoRealizace.psc.value == data.bydliste.psc.value &&
+				data.mistoRealizace.psc.value != '' &&
+				data.mistoRealizace.ulice.value == data.bydliste.ulice.value &&
+				data.mistoRealizace.ulice.value != ''
+			) {
+				data.mistoRealizace.jakoBydliste.value = true;
+				data.mistoRealizace.obec.updateText('');
+				data.mistoRealizace.psc.value = '';
+				data.mistoRealizace.ulice.value = '';
+			}
 		}
-	}
-	$: if (mode != 'loading') {
-		if (data.ir.typ.value.second == p`RTC`) {
-			data.tc.typ.value = 'airToWater';
+	});
+	$effect(() => {
+		data
+		if (mode != 'loading') {
+			if (!data.tc.model.moznosti(data).includes(data.tc.model.value ?? '')) {
+				data.tc.model.value = null;
+			}
 		}
-	}
-	$: if (mode != 'loading' && mode != 'edit') {
-		storedData.set(dataToRawData(data));
-	}
-	$: typBOXu = typBOX(data.ir.cisloBOX.value);
+	});
+	$effect(() => {
+		data
+		if (mode != 'loading') {
+			if (data.ir.typ.value.second == p`RTC`) {
+				data.tc.typ.value = 'airToWater';
+			}
+		}
+	});
+	$effect(() => {
+		if (mode != 'loading' && mode != 'edit') {
+			storedData.set(dataToRawData(data));
+		}
+	});
+	let typBOXu = $derived(typBOX(data.ir.cisloBOX.value));
 
 	const filter = writable('');
-	$: filtered = filteredCompanies(filter);
+	let filtered = $derived(filteredCompanies(filter));
 
-	$: chosen =
-		mode != 'loading'
+	let chosen = $derived(
+		data && mode != 'loading'
 			? {
 					assemblyCompany: nazevFirmy(data.montazka.ico.value),
 					commissioningCompany: nazevFirmy(data.uvedeni.ico.value)
 				}
-			: undefined;
+			: undefined
+	);
 
-	$: list =
-		mode != 'loading'
+	let list = $derived(
+		data && mode != 'loading'
 			? (Object.values(data) as Data[keyof Data][]).flatMap(
 					(obj) => Object.values(obj) as v.Vec<Data, any>[]
 				)
-			: [];
+			: []
+	);
 
-	let vysledek = {
+	let vysledek = $state({
 		text: '',
 		red: false,
 		load: false
-	};
-	let doNotSend = false;
+	});
+	let doNotSend = $state(false);
 </script>
 
 <div class="d-flex flex-row flex-wrap">
@@ -146,7 +168,7 @@
 		{#if mode == 'create' || mode == 'createStored'}
 			<button
 				class="btn"
-				on:click={() => {
+				onclick={() => {
 					$storedData = undefined;
 					window.location.reload();
 				}}
@@ -172,7 +194,7 @@
 	</div>
 </div>
 
-{#each list as vec}
+{#each list as vec, i}
 	{#if vec === data.montazka.ico && vec.zobrazit(data) && vec instanceof v.Pisatkova}
 		{#if $companies.assemblyCompanies.length > 0}
 			<VybiratkoFirmy
@@ -187,7 +209,12 @@
 				{t}
 			/>
 		{/if}
-		<Pisatko bind:vec {t} {data} disabled={vec == data.ir.cislo && mode == 'edit'} />
+		<Pisatko
+			bind:vec={list[i] as v.Pisatkova<Data>}
+			{t}
+			{data}
+			disabled={vec == data.ir.cislo && mode == 'edit'}
+		/>
 	{:else if vec === data.uvedeni.ico && vec.zobrazit(data) && vec instanceof v.Pisatkova}
 		{#if $companies.commissioningCompanies.length > 0}
 			<VybiratkoFirmy
@@ -202,7 +229,12 @@
 				{t}
 			/>
 		{/if}
-		<Pisatko bind:vec {t} {data} disabled={vec == data.ir.cislo && mode == 'edit'} />
+		<Pisatko
+			bind:vec={list[i] as v.Pisatkova<Data>}
+			{t}
+			{data}
+			disabled={vec == data.ir.cislo && mode == 'edit'}
+		/>
 	{:else if vec === data.tc.cislo && vec.zobrazit(data)}
 		<Scanner
 			bind:vec={data.tc.cislo}
@@ -216,19 +248,26 @@
 	{:else if vec instanceof v.Textova && vec.zobrazit(data)}
 		<p>{t.get(vec.nazev(data))}</p>
 	{:else if vec instanceof v.Pisatkova && vec.zobrazit(data)}
-		<p><Pisatko bind:vec {t} {data} disabled={vec == data.ir.cislo && mode == 'edit'} /></p>
+		<p>
+			<Pisatko
+				bind:vec={list[i] as v.Pisatkova<Data>}
+				{t}
+				{data}
+				disabled={vec == data.ir.cislo && mode == 'edit'}
+			/>
+		</p>
 	{:else if vec instanceof v.DvojVybiratkova && vec.zobrazit(data)}
-		<p><DvojVybiratko bind:vec {t} {data} /></p>
+		<p><DvojVybiratko bind:vec={list[i] as v.DvojVybiratkova<Data>} {t} {data} /></p>
 	{:else if vec instanceof v.Vybiratkova && vec.zobrazit(data)}
-		<p><Vybiratko bind:vec {t} {data} /></p>
+		<p><Vybiratko bind:vec={list[i] as v.Vybiratkova<Data>} {t} {data} /></p>
 	{:else if vec instanceof v.Radiova && vec.zobrazit(data)}
-		<p><Radio bind:vec {t} {data} /></p>
+		<p><Radio bind:vec={list[i] as v.Radiova<Data>} {t} {data} /></p>
 	{:else if vec instanceof v.Prepinatkova && vec.zobrazit(data)}
-		<p><Prepinatko bind:vec {t} {data} /></p>
+		<p><Prepinatko bind:vec={list[i] as v.Prepinatkova<Data>} {t} {data} /></p>
 	{:else if vec instanceof v.MultiZaskrtavatkova && vec.zobrazit(data)}
-		<p><MultiZaskrtavatko {t} bind:vec {data} /></p>
+		<p><MultiZaskrtavatko {t} bind:vec={list[i] as v.MultiZaskrtavatkova<Data>} {data} /></p>
 	{:else if vec instanceof v.Zaskrtavatkova && vec.zobrazit(data)}
-		<p><Zaskrtavatko {t} bind:vec {data} /></p>
+		<p><Zaskrtavatko {t} bind:vec={list[i] as v.Zaskrtavatkova<Data>} {data} /></p>
 	{/if}
 	{#if vec == data.ir.cisloBOX && vec.zobrazit(data) && typBOXu}
 		<p>Rozpoznáno: {typBOXu}</p>
@@ -271,7 +310,7 @@
 				id="odeslat"
 				type="button"
 				class="btn btn-success"
-				on:click={() =>
+				onclick={() =>
 					odeslat(
 						data,
 						(v) => (vysledek = v),
@@ -288,7 +327,7 @@
 			</button>
 		{/if}
 		{#if mode == 'edit'}
-			<button type="button" class="btn btn-outline-secondary ms-2" on:click={() => history.back()}>
+			<button type="button" class="btn btn-outline-secondary ms-2" onclick={() => history.back()}>
 				{t.back}
 			</button>
 		{/if}
@@ -296,7 +335,7 @@
 
 	<div class="d-inline-flex align-content-center text-break mt-2 mt-sm-0">
 		{#if vysledek.load}
-			<div class="spinner-border text-danger ms-2" />
+			<div class="spinner-border text-danger ms-2"></div>
 		{/if}
 		<p class:text-danger={vysledek.red} class="ms-2 my-auto">{@html vysledek.text}</p>
 	</div>
